+++ src/commands/AboutCommand.ts
@@ -3,7 +3,7 @@ import { MessageEmbed, version } from "discord.js";
 import { uptime as osUptime } from "os";
 import path from "path";
 import { formatMS } from "../utils/formatMS";
-import type Jukebox from "../structures/Jukebox";
+import type Jukebox from "../structures/Disc_11";
 import type { IMessage } from "../../typings";
 
 export default class AboutCommand extends BaseCommand {
